Fixed a bug in the conversion from DateTime to CIMDateTime that produced wrong GMT offsets in some cases.
Also removed the function DateTime::getGMTOffset().  Not only was it
implemented wrong, but its entire concept was flawed:
- It is a static function and assumes that the UTC offset is fixed for a
  timezone, when in reality the offset varies depending on whether or not
  daylight savings time is in effect.
- It returns the offset as a number of hours, but there are several time
  zones (including India Standard Time and one of the Canadian time zones)
  that are offset by an odd multiple of 30 minutes.
Furthermore, the only place that OpenWBEM used this function was in the
conversion from DateTime to CIMDateTime.

If there's anybody out there who was using DateTime::getGMTOffset(), I created
a replacement function: the non-static function DateTime::toLocal().