Converting the (fairly useless) IPCIO exceptions into more useful exceptions.
Aborting the provider (debug mode only) if the PrivilegeMonitor says the
provider requested something that it was not allowed to do.  This will aid in
the debugging process, as a useful core file will result (if the core pattern
allows core creation).

Tagging the connection with the monitor as invalid whenever an error condition
results from communication with the monitor.  The monitor is designed to shut
down in cases like that and any further attempts to make requests would result
in more IPCIO errors.  By tagging the connection as invalid, a useful exception
can politely inform the user (who must have somehow ignored a prior exception)
that there is no longer a connection to the monitor.

Inserting all of the try/catch blocks caused a lot of indenting changes to
happen, so a diff -b on this set of changes will produce the most useful
results.