Removing the PRIVOP_BODY and PRIVOP_FINISH macros.
Fixing a few places where the proper data was not being sent back to the
privilege manager in the case of errors (missing error codes), which should
have resulted in binary corruption of the error data.  These are now all
reporting the error data through a common function, so there is only one
point of modification, should the protocol change.

Attaching error codes to (almost) all places where exceptions are thrown.
This will be used by the privilege manager to determine if the exception was
critical or not.