Fixing a bug in the multireq indications where the trailing \r that we include in the XML sent on the remote end can cause an exception when parsing, as it shows up as data.
This fix will still be wrong if things change so we expect further data to 
exist in stream sent to the xml parser -- it will eat too much.  In its 
current state, the multireq parsing will continue to eat simplereq nodes, as 
long as they exist and stop after no more are available.  This doesn't make
any attempt to verify that there are no unexpected xml nodes after the group
of simplereq nodes, mainly because the parser.getNextId() function keeps 
eating nodes until it finds one it likes or until it runs out of nodes.